Add case-insensitive HTTP server and update README
This commit adds a new Python script, , which implements a simple HTTP server that is case-insensitive to file paths. This is primarily intended for use with games that were developed for Windows and have issues with case-sensitive file paths on Linux.

The README was also updated to reflect the addition of this new script and its purpose.

Signed-off-by: Li Hua <lihua@email.com>